レビューされた内容の修正
