+++ app/Http/Controllers/Api/CourseController.php
@@ -11,6 +11,7 @@ use App\Http\Resources\CourseShowResource;
 use App\Http\Resources\CourseProgressResource;
 use App\Model\Attendance;
 use App\Model\Course;
+use App\Model\LessonAttendance;
 
 class CourseController extends Controller
 {
@@ -83,6 +84,105 @@ class CourseController extends Controller
         ])
         ->firstOrFail();
 
-        return new CourseProgressResource($attendance);
+        $progressDate = [
+            'getCompletedChaptersCount' => $this->getCompletedChaptersCount($attendance),
+            'getTotalChaptersCount' => $this->getTotalChaptersCount($attendance),
+            'getCompletedLessonsCount' => $this->getCompletedLessonsCount($attendance),
+            'getTotalLessonsCount' => $this->getTotalLessonsCount($attendance),
+            'getYoungestUnCompletedLessonId' => $this->getYoungestUnCompletedLessonId($attendance)
+        ];
+
+        return new CourseProgressResource($attendance, $progressDate);
+    }
+
+        /**
+     * 完了済みのチャプター数を取得する
+     *
+     * @param Attendance $attendance
+     * @return int
+     */
+    private function getCompletedChaptersCount($attendance)
+    {
+        return $attendance->course->chapters->filter(function ($chapter) use ($attendance) {
+            $isCompleted = false;
+            // 全てのレッスンが完了済みかどうかをチェック
+            $chapter->lessons->each(function ($lesson) use ($attendance, &$isCompleted) {
+                $lessonAttendance = $attendance->lessonAttendances->where('lesson_id', $lesson->id)->first();
+                if ($lessonAttendance->status !== LessonAttendance::STATUS_COMPLETED_ATTENDANCE) {
+                    $isCompleted = false;
+                    return false;
+                }
+                $isCompleted = true;
+            });
+            return $isCompleted;
+        })->count();        
+    }
+
+    /**
+     * チャプター合計を取得する
+     *
+     * @param Attendance $attendance
+     * @return int
+     */
+    private function getTotalChaptersCount($attendance)
+    {
+        return $attendance->course->chapters->count();
+    }
+
+    /**
+     * 完了済みのレッスン数を取得する
+     *
+     * @param Attendance $attendance
+     * @return int
+     */
+    private function getCompletedLessonsCount($attendance)
+    {
+        return $attendance->lessonAttendances->filter(function ($lessonAttendance) {
+            return $lessonAttendance->status === LessonAttendance::STATUS_COMPLETED_ATTENDANCE;
+        })->count();
+    }
+
+    /**
+     * レッスン合計を取得する
+     *
+     * @param Attendance $attendance
+     * @return int
+     */
+    private function getTotalLessonsCount($attendance)
+    {
+        $totalLessonsCount = 0;
+        foreach ($attendance->course->chapters as $chapter) {
+            $lessonCount = $chapter->lessons->count();
+            $totalLessonsCount += $lessonCount; 
+        }
+        return $totalLessonsCount;
+    }
+
+    /**
+     * 続きのレッスンIDを取得する
+     *
+     * @param Attendance $attendance
+     * @return int|null
+     */
+    private function getYoungestUnCompletedLessonId($attendance)
+    {
+        // IDが最も若い未完了のチャプターの内、IDが最も若い未完了のレッスン
+        $youngestUnCompletedLessonId = null;
+        $attendance->course->chapters->each(function ($chapter) use ($attendance, &$youngestUnCompletedLessonId) {
+            $chapter->lessons->each(function ($lesson) use ($attendance, &$youngestUnCompletedLessonId) {
+                $lessonAttendance = $attendance->lessonAttendances->where('lesson_id', $lesson->id)->first();
+
+                if ($lessonAttendance->status !== LessonAttendance::STATUS_COMPLETED_ATTENDANCE) {
+                    if ($youngestUnCompletedLessonId === null) {
+                        $youngestUnCompletedLessonId = $lesson->id;
+                        return;
+                    }
+                    if ($youngestUnCompletedLessonId > $lesson->id) {
+                        $youngestUnCompletedLessonId = $lesson->id;
+                    }
+                }
+            });
+        });
+        return $youngestUnCompletedLessonId;
     }
 }
